main: Change stat types to gint64
The JSON library will read and write the values as `gint64`, this means
that on a 32 bit system, we are incorrectly storing the stat values.

e.g. writeSectors is an unsigned long, by reading with `strtoul` and then
casting to `gint64` we avoid it becoming negative in a cast from
unsigned to signed, or dropping bits (if we were instead to use the
`int` type)

totalBytesWritten does not need to be a `double` and was also changed to a
`gint64`

Closes #54

Signed-off-by: Benjamin Braddock <bbraddock@thegoodpenguin.co.uk>