main: Fix stats updating
The current implementation doesn't account for either the device being
removed and reconnected, or the host rebooting.

We now have two stats structs in the `jsonDeviceEntry`, one to keep
track of the current stats and one to store to / read from the json
file. This way we can update the stats without resetting the
previous data.

The disk sequence is also read in order to detect when the stats have
been reset due to to the card being removed and then reconnected
inbetween updates. This will only keep track of writes after the last
reconnect, and any writes that occured before that (inbetween the
KrillKounter update period) will not be tracked. This of course can be
solved by increasing the update rate.

Closes #38

Signed-off-by: Benjamin Braddock <bbraddock@thegoodpenguin.co.uk>