main: add multi device monitoring
This uses the newly added JSON config parser (#77) to declare the
devices to monitor. Currently we have an array of device paths, and
a single stats file output, and single update rate.

The config file is in the format

```JSON
{
    "devices": [
        "/dev/sda",
        "/dev/mmcblk0"
    ],
    "updateRate": 3600,
    "statsFilePath": "/usr/share/KrillKounter/stats.json"
}
```

Modified the necessary functions to iterate over all devices in the
array instead of working on a single global variable.

Closes #46

Signed-off-by: Benjamin Braddock <bbraddock@thegoodpenguin.co.uk>