Added array parsing fallback for version < 1.8.0
For json-glib versions <1.8.0 (e.g. on kirkstone) the
`json_reader_get_current_node` function does not exist, and as such we
are not able to use the array functions with the node pointer. We check
the library version to decide which method to use; should the old interface be
deprecated or the flexibility of the newer interface be desired for new
features, it will be helpful to have both versions.

Closes #85

Signed-off-by: Benjamin Braddock <bbraddock@thegoodpenguin.co.uk>