main: write counters on service stopping
This commit makes the stats to be written on daemon termination.

The onExit() function is not called when the daemon is terminated,
but only when exit(). Therefore, calling updateAllDeviceStats()
from onExit() has no effect when the daemon terminates. Instead,
this commit adds a signal handler which exits the pLoop when
terminating the daemon, and adds a call to updateAllDeviceStats()
after the loop, causing the stats to be updated after the
terminating signal is received. The custom handler is unregistered
after exiting the pLoop, so the default handler can continue
servicing the signal after updating the statistics.

Signed-off-by: Noe Galea <ngalea@thegoodpenguin.co.uk>