Add compile time checks for std::format
std::format isn't supported on older versions of gcc, even though they do
support C++20. It looks like it was only implemented starting in gcc 14.
Thus, add a compile time check to see if std::format is supported and
add a different implementation if it isn't supported. Tested on gcc
11.4.0, where std::format isn't implemented and gcc 14 is implemented.